Trying to create an HTML word clock.  I think I'm most of the way done. I finally got the letters in an order that makes sense for every time possibility.
The spaces in between the time words spell out IAN ALEXANDER
FANG which all have the id="ian".  Except the second A which has
id="ian_a" since it is also used with the quarter.

I still need to figure out how to clear the styles before writing the next
time.

This is a module that is within the magic mirror project.  It will be one
of the things that can be brought up when controlling the mirror through
Alexa.

timeSplit.js is a prototyping file.