Updated the word clock with new javascript skills I have recently leanred.
I was having a bunch of problems with setTimeout.  Now I know that only
the function in setTimeout will run after the timeout stated.  Any code
that comes after the setTimeout will be run right away.  I guess
setTimeout() creates a call back that is called after the timeout expires.
I found it strange that I couldn't use a direct function call
setNewTime(getCurrentTime()) in the setTimeout.  It would make that call
immediately.  I had to create an anonymous function in order for
setTimeout() to behave as I expected it would.

When I was playing with the setTimeout(), I somehow got into a state where
the mouseup eventlistener, would call buttonPress, which would then call
setNewTime(getCurrentTime()) immediately (since it wasn't in an anonymous
function) and would never be updated with the time even though it was
calling decodeTime and time.minute % 5 === 1 and timeUpdate were true.
Not sure what was happening there, but once I got the setTimeout() to work
correctly, I didn't see the issue again.

Adding and removing classes seems a lot cleaner than directly setting the
style on each span.  If I ever want to change the colors or style, I do it
for everything by changing it in the css instead of updating consts in the
javascript.

I think there's a better way to do decodeTime() since right now I'm
setting timeUpdated to false 60 times.  I can't think of it now, but it's
better than calling setNewTime() 60 times.