Review from Craig Jennings: Overall it looks really good. Below are just nit-picks, nothing huge. Hope the new year has started well for you!
wordclock.js
  L46-49: You can just instantiate `ids` with "it" and "is" immediately
  instead of "push"ing them onto the array.

      In `clearStyle`, `for` loops are less common in javascript. An
      array's `forEach` method is commonly used. This means you could
      rewrite those to be `elements.forEach((el) =>
      el.classList.remove('work-on'));`

      wordclock.html
        Dont' forget `<!DOCTYPE html>` at the top for properly formed html
	  The button "name" attribute is rarely used. Usually you assign
	  it an "id" (there's also a missing "=" for the name attribute)

-----------------------------

I was thinking that pushing everything in to the array and then calling
classList.add() once would be easier to read, but it doesn't really matter
to me either way.  Is it also an efficiency thing to immediately
instantiate those id's?

I took away all the for loops and replaced them with the forEach.  I also
realized that I could have use multiple items in query selector all which
simplified the code some more.